Fix lua issues

+++ InjectedTimer/ScriptEngine.cs
@@ -14,11 +14,11 @@ namespace InjectedTimer
 
             luaState.RegisterFunction("module_base_address", ((Func<string, IntPtr>)ScriptFunction.GetModuleAddress).GetMethodInfo());
             luaState.RegisterFunction("debug_message", ((Action<string>)ScriptFunction.DebugMessage).GetMethodInfo());
-            luaState.RegisterFunction("debug_message_address", ((Action<IntPtr>)ScriptFunction.DebugMessageAddress).GetMethodInfo());
+            luaState.RegisterFunction("debug_message_address", ((Action<long>)ScriptFunction.DebugMessageAddress).GetMethodInfo());
             luaState.RegisterFunction("timer_set_loading", ((Action)ScriptFunction.SetLoading).GetMethodInfo());
             luaState.RegisterFunction("timer_set_not_loading", ((Action)ScriptFunction.SetNotLoading).GetMethodInfo());
 
-            luaState.RegisterFunction("read_int32", ((Func<IntPtr, int>)ScriptFunction.ReadInt32).GetMethodInfo());
+            luaState.RegisterFunction("read_int32", ((Func<long, int>)ScriptFunction.ReadInt32).GetMethodInfo());
 
             luaState.DoString(script);
             luaState.DoString("on_load()");
